Improve station selection on adding a stop to a journey
This commit adds the ability to specify the route of a journey's
stop by either specifying only the next arrival station or by specifying
both the departure station and the arrival station.
Previously, it was only possible to specify the next arrival station
and use the previous arrival station as departure station.